requirements.txt
django==3.2.8
psycopg2==2.9.1
python-decouple==3.6
djangorestframework==3.13.1
requests==2.27.1
Unipath==1.1